+++ qetpy/core/didv/_didv_plotting.py
@@ -1,5 +1,6 @@
 import numpy as np
 import matplotlib.pyplot as plt
+
 from ._base_didv import squarewaveresponse, complexadmittance
