+++ qetpy/utils/_utils.py
@@ -11,7 +11,6 @@ __all__ = [
     "get_offset_from_muon",
     "powertrace_simple",
     "energy_absorbed",
-    "stdcomplex",
     "slope",
     "fill_negatives",
     "shift",
@@ -142,34 +141,6 @@ def make_decreasing(y, x=None):
 
     return out
 
-def stdcomplex(x, axis=0):
-    """
-    Function to return complex standard deviation (individually
-    computed for real and imaginary components) for an array of complex
-    values.
-
-    Parameters
-    ----------
-    x : ndarray
-        An array of complex values from which we want the complex
-        standard deviation.
-    axis : int, optional
-        Which axis to take the standard deviation of (should be used if
-        the dimension of the array is greater than 1).
-
-    Returns
-    -------
-    std_complex : ndarray
-        The complex standard deviation of the inputted array, along the
-        specified axis.
-
-    """
-
-    rstd = np.std(x.real, axis=axis)
-    istd = np.std(x.imag, axis=axis)
-    std_complex = rstd + 1.0j * istd
-
-    return std_complex
 
 def calc_offset(x, fs=1.0, sgfreq=100.0, is_didv=False):
     """
